pyproject.toml
[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[project]
name = "opanhome-hub"
version = "0.1.0"
description = "ESPHome-native voice hub scaffold for opanhome"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "aioesphomeapi==42.7.0",
  "httpx>=0.28,<1",
  "python-dotenv>=1,<2",
  "typer>=0.12,<1",
  "websockets>=15,<16",
]

[project.scripts]
hub = "hub.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["hub"]

[dependency-groups]
dev = [
  "pytest>=8,<9",
]
